pyproject.toml
[project]
name = "updates2mqtt"
readme = "README.md"
description = "System update and docker image notification and execution over MQTT"
authors = [
    { name = "jey burrows", email = "jrb@rhizomatics.org.uk" }
]

requires-python = ">=3.13"
version = "1.5.1"
license="Apache-2.0"
keywords=["mqtt", "docker", "updates", "automation","home-assistant","homeassistant","selfhosting"]

dependencies = [
  "docker>=7.1.0",
  "paho-mqtt>=2.1.0",
  "omegaconf>=2.3.0",
  "structlog>=25.4.0",
  "rich>=14.0.0",
  "httpx>=0.28.1",
  "hishel[httpx]>=0.1.4",
  "usingversion>=0.1.2",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: Other/Proprietary License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Topic :: Home Automation",
    "Topic :: System :: Systems Administration",
    "Topic :: System :: Monitoring",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Typing :: Typed",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.13"
]

[project.scripts]
updates2mqtt = "updates2mqtt.app:run"

[tool.uv]
compile-bytecode = true
managed = true

[dependency-groups]
dev = [
    "ruff>=0.13.0",
    "pre-commit>=4.4.0",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.0",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-httpx",
    "pytest-mqtt>=0.6.0",
    "pytest-subprocess>=1.5.3",
    "coverage",
    "icdiff",
    "genbadge[all]"
]
mkdocs=[
    "mkdocs",
    "mkdocs-material",
    "mkdocs-minify-plugin",
    "mkdocs-autorefs",
    "mkdocs-pagetree-plugin",
    "mkdocs-coverage",
    "mkdocstrings[python]",
    "pymdown-extensions",
    "mkdocs-git-revision-date-localized-plugin",
    "mkdocs-meta-descriptions-plugin",
    "pngquant"
]

[build-system]
build-backend = "hatchling.build"
requires=[
  "hatchling"
]

[tool.bandit]
exclude_dirs = ["tests"]
skips=["B404", #import subprocess
"B602" #popen with shell=True
]

[tool.ruff]
line-length = 128
indent-width = 4
target-version = "py313"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "htmlcov",
    ".mypy_cache",
    ".pytest_cache",
    ".github",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
select = ["E", "C","R", "F", "B", "W", "I", "C90", "N", "D", "UP", "YTT", 
    "RUF","PGH","PT","T20","DTZ","TD","PTH","ARG","INT","TCH","TID","SIM","PERF","ASYNC","ANN","S","A","COM","C4"]
ignore = [
    "W191","E111","E114","E117","D206","D300","Q000","Q001","COM812","COM819","ISC001","ISC002", # std formatter conflicts
    "D203","D213", # other formatter clashes
    "TD002","TD003","TD005",
    "D103","D100","D101","D102","D104","D107","D415","D400","D415","ANN401",
    "S101","C901","FURB113",
    "S602","S607","S404" ] # subprocess usage
unfixable = ["B"]


[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = [
    "tests",
]
norecursedirs = [
    ".git",
    "templates",
]

pythonpath = [
  "."
]
addopts = "--junitxml=junit/test-results.xml --cov --cov-report=lcov:lcov.info --cov-report=xml --cov-report=html --cov-report=term --import-mode=importlib"

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
fail_under=0

[tool.coverage.xml]
output="cov.xml"

[tool.coverage.html]
directory="htmlcov"

[tool.mypy]
mypy_path=["./src","./tests"]

[project.urls]
Homepage="https://updates2mqtt.rhizomatics.org.uk"
Repository="https://github.com/rhizomatics/updates2mqtt"
Documentation="https://updates2mqtt.rhizomatics.org.uk"
Issues="https://github.com/rhizomatics/updates2mqtt/issues"
Changelog="https://github.com/rhizomatics/updates2mqtt/blob/main/CHANGELOG.md"